Add a gemspec file so that this can build a gem
I had to move the git-rank executable to the bin dir so the gem would
work properly.  Now you can run `gem build git-rank.gemspec' and then
install the resulting gem to use this software.  It's still better to
run from source since it's not to a point that I would consider
releasable to many others, but it's a start.

The README has been renamed so that it renders as markdown on Github,
and also instructions on how to build and install as a gem have been
added.