Add a Rake task to generate a standalone script
Even though the code is easier to manage and test as something like a
Ruby project, it's harder to install and use.  The rake task added
`rake bundle` will output an executable, standalone script called
git-rank in the project root.

The README install instructions have also been updated to give this as
an options, and provide better documentation to the other installation
options.